alertInvalidAuth added

+++ backEnd/src/controllers/todoController.js
@@ -1,37 +1,68 @@
 import mongoose from "mongoose";
 import { TaskModel } from "../model/todoModel.js";
+import authRequest from "../auth/auth.js";
+
 
 export const getAllTasks = async (req, res) => {
   try {
-    const tasks = await TaskModel.find();
-    res.status(200).json(tasks);
+    const validAuthRequest = authRequest(req);
+    if (validAuthRequest) {
+      const tasks = await TaskModel.find();
+      res.status(200).json(tasks);
+    }
   } catch (error) {
     res.status(404).json({ messageError: error.message });
   }
 };
 
 export const getTasksByUsername = async (req, res) => {
+  const { username } = JSON.parse(req.params.optionGetTasks);
+
   try {
-    const { username } = req.params;
-    const tasks = await TaskModel.find({user:username});
-    res.status(200).json(tasks);
+    const validAuthRequest = authRequest(req, res);
+    if (validAuthRequest) {
+      const tasks = await TaskModel.find({ user: username });
+      res.status(200).json(tasks);
+    }
+  } catch (error) {
+    res.status(404).json({ messageError: error.message });
+  }
+};
+
+export const getStateTasksByUsername = async (req, res) => {
+  const { isCompleted, username } = JSON.parse(req.params.optionGetTasks);
+  try {
+    const validAuthRequest = authRequest(req);
+    if (validAuthRequest) {
+      const tasksStateSelected = await TaskModel.find({
+        isCompleted: isCompleted,
+        user: username,
+      });
+
+      res.status(200).json(tasksStateSelected);
+    }
   } catch (error) {
     res.status(404).json({ messageError: error.message });
   }
 };
 
 export const getTaskById = async (req, res) => {
+  const { id } = JSON.parse(req.params.optionGetTasks);
+
   try {
-    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
-      throw new Error("Id not valid");
-    }
-    const taskFoundById = await TaskModel.findById(req.params.id);
+    const validAuthRequest = authRequest(req);
+    if (validAuthRequest) {
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw new Error("Id not valid");
+      }
+      const taskFoundById = await TaskModel.findById(id);
 
-    if (!taskFoundById) {
-      throw new Error("Task not found");
-    }
+      if (!taskFoundById) {
+        throw new Error("Task not found");
+      }
 
-    res.status(200).json(taskFoundById);
+      res.status(200).json(taskFoundById);
+    }
   } catch (error) {
     res.status(404).json({ messageError: error.message });
   }
@@ -41,14 +72,18 @@ export const createTask = async (req, res) => {
   const task = req.body;
 
   try {
-    const taskExisting = await findTaskExisting(task);
+    const validAuthRequest = authRequest(req);
 
-    if (taskExisting) {
-      throw new Error("This task already exists");
+    if (validAuthRequest) {
+      const taskExisting = await findTaskExisting(task);
+
+      if (taskExisting) {
+        throw new Error("This task already exists");
+      }
+      const taskCreated = await TaskModel.create(task);
+
+      res.status(201).json(taskCreated);
     }
-    const taskCreated = await TaskModel.create(task);
-
-    res.status(201).json(taskCreated);
   } catch (error) {
     res.status(502).json({ messageError: error.message });
   }
@@ -71,17 +106,21 @@ const findTaskExisting = async (task) => {
 
 export const updateTask = async (req, res) => {
   try {
-    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
-      throw new Error("Id not valid");
+    const validAuthRequest = authRequest(req);
+
+    if (validAuthRequest) {
+      if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        throw new Error("Id not valid");
+      }
+
+      const taskUpdated = await TaskModel.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        { new: true }
+      );
+
+      res.status(200).json(taskUpdated);
     }
-
-    const taskUpdated = await TaskModel.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      { new: true }
-    );
-
-    res.status(200).json(taskUpdated);
   } catch (error) {
     res.status(404).json({ messageError: error.message });
   }
@@ -89,14 +128,18 @@ export const updateTask = async (req, res) => {
 
 export const deleteTask = async (req, res) => {
   try {
-    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
-      throw new Error("Id not valid");
-    }
+    const validAuthRequest = authRequest(req);
 
-    await TaskModel.findByIdAndDelete(req.params.id);
-    res
-      .status(201)
-      .json({ message: `Task ${req.params.id} deleted succesfully` });
+    if (validAuthRequest) {
+      if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        throw new Error("Id not valid");
+      }
+
+      await TaskModel.findByIdAndDelete(req.params.id);
+      res
+        .status(201)
+        .json({ message: `Task ${req.params.id} deleted succesfully` });
+    }
   } catch (error) {
     res.status(404).json({ messageError: error.message });
   }
